refactor:change user repository file name and location
